requirements.txt
google-adk>=1.0.0
google-cloud-speech>=2.0.0
google-cloud-texttospeech>=2.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
